:recycle: api: move flash assessment simulator related models to src
