+++ api/lib/domain/models/index.js
@@ -6,9 +6,9 @@ import { AnswerCollectionForScoring } from './AnswerCollectionForScoring.js';
 import { AnswerStatus } from './AnswerStatus.js';
 import { Area } from './Area.js';
 import { Assessment } from '../../../src/shared/domain/models/Assessment.js';
-import { FlashAssessmentAlgorithm } from './flash-assessment-algorithm/FlashAssessmentAlgorithm.js';
+import { FlashAssessmentAlgorithm } from '../../../src/certification/flash-certification/domain/model/FlashAssessmentAlgorithm.js';
 import { AssessmentResult } from './AssessmentResult.js';
-import { AssessmentSimulator } from './AssessmentSimulator.js';
+import { AssessmentSimulator } from '../../../src/certification/flash-certification/domain/model/AssessmentSimulator.js';
 import { Authentication } from './Authentication.js';
 import { AuthenticationMethod } from './AuthenticationMethod.js';
 import { AuthenticationSessionContent } from './AuthenticationSessionContent.js';
