+++ api/src/certification/flash-certification/domain/model/FlashAssessmentAlgorithmForcedCompetencesRule.js
