+++ api/src/certification/flash-certification/domain/model/FlashAssessmentAlgorithmOneQuestionPerTubeRule.js
