+++ api/src/certification/flash-certification/domain/model/FlashAssessmentAlgorithmPassageByAllCompetencesRule.js
