+++ api/src/certification/flash-certification/domain/model/FlashAssessmentSuccessRateHandlerFixedStrategy.js
