+++ api/tests/certification/flash-certification/unit/domain/models/FlashAssessmentAlgorithmPassageByAllCompetencesRule_test.js
@@ -1,5 +1,5 @@
-import { FlashAssessmentAlgorithmPassageByAllCompetencesRule } from '../../../../../lib/domain/models/flash-assessment-algorithm/FlashAssessmentAlgorithmPassageByAllCompetencesRule.js';
-import { domainBuilder, expect } from '../../../../test-helper.js';
+import { FlashAssessmentAlgorithmPassageByAllCompetencesRule } from '../../../../../../src/certification/flash-certification/domain/model/FlashAssessmentAlgorithmPassageByAllCompetencesRule.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | FlashAssessmentAlgorithm | FlashAssessmentAlgorithmPassageByAllCompetencesRule', function () {
   describe('#isApplicable', function () {
