+++ api/tests/school/unit/application/assessments/route_test.js
@@ -1,7 +1,7 @@
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 import { assessmentController } from '../../../../../src/school/application/assessments/assessment-controller.js';
 import * as moduleUnderTest from '../../../../../src/school/application/assessments/assessment-route.js';
-import { AssessmentEndedError } from '../../../../../lib/domain/errors.js';
+import { AssessmentEndedError } from '../../../../../src/shared/domain/errors.js';
 
 describe('Unit | Application | Router | assessment-router', function () {
   describe('GET /api/pix1d/assessments/{id}/next', function () {
